Eliminate code duplication on performance report
