+++ src/components/performance-review-dashboard.tsx
@@ -10,6 +10,8 @@ import {
   subWeeks,
 } from "date-fns";
 import type { StandupResponse } from "@/lib/matt-api";
+import { UserDetailedView } from "./user-detailed-view";
+import { MobileModal } from "./mobile-modal";
 
 interface PerformanceReviewDashboardProps {
   orgName: string;
@@ -386,629 +388,16 @@ export function PerformanceReviewDashboard({
 
         {/* Detailed View Section - Hidden on mobile */}
         <div className="hidden lg:flex lg:flex-col bg-white rounded-lg border border-gray-200 overflow-hidden">
-          {selectedUser ? (
-            <>
-              {/* User header - fixed */}
-              <div className="flex-shrink-0 p-6 border-b border-gray-200">
-                <div className="flex items-center gap-3">
-                  <img
-                    src={selectedUser.avatar_url}
-                    alt={selectedUser.username}
-                    className="w-12 h-12 rounded-full"
-                  />
-                  <div>
-                    <h3 className="font-semibold text-gray-900">
-                      {selectedUser.name || selectedUser.username}
-                    </h3>
-                    <p className="text-sm text-gray-500">
-                      @{selectedUser.username}
-                    </p>
-                  </div>
-                </div>
-              </div>
-
-              {/* Scrollable content */}
-              <div className="flex-1 overflow-y-auto p-6">
-                <div className="space-y-4">
-                  {/* Metrics */}
-                  <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-                    <div className="bg-gray-50 p-2 rounded-lg text-center">
-                      <div className="font-semibold text-green-600">
-                        {selectedUser.totalManHoursMin}-
-                        {selectedUser.totalManHoursMax}h
-                      </div>
-                      <div className="text-sm text-gray-500">Man-Hours</div>
-                    </div>
-                    <div className="bg-gray-50 p-2 rounded-lg text-center">
-                      <div className="font-semibold text-blue-600">
-                        {selectedUser.avgManHours.toFixed(1)}h
-                      </div>
-                      <div className="text-sm text-gray-500">Average</div>
-                    </div>
-                    <div className="bg-gray-50 p-2 rounded-lg text-center">
-                      <div className="font-semibold text-purple-600">
-                        {selectedUser.totalActivities}
-                      </div>
-                      <div className="text-sm text-gray-500">Activities</div>
-                    </div>
-                    <div className="bg-gray-50 p-2 rounded-lg text-center">
-                      <div className="font-semibold text-orange-600">
-                        {selectedUser.activeDays}
-                      </div>
-                      <div className="text-sm text-gray-500">Days</div>
-                    </div>
-                  </div>
-
-                  {/* Activity breakdown */}
-                  <div className="flex items-center justify-center gap-4 text-sm text-gray-600 bg-gray-50 p-2 rounded-lg">
-                    <span className="flex items-center gap-1">
-                      <svg
-                        className="w-3.5 h-3.5 text-green-600"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          d="M2 5a2 2 0 012-2h12a2 2 0 012 2v10a2 2 0 01-2 2H4a2 2 0 01-2-2V5zm3.293 1.293a1 1 0 011.414 0l3 3a1 1 0 010 1.414l-3 3a1 1 0 01-1.414-1.414L7.586 10 5.293 7.707a1 1 0 010-1.414zM11 12a1 1 0 100 2h3a1 1 0 100-2h-3z"
-                          clipRule="evenodd"
-                        />
-                      </svg>
-                      {selectedUser.totalCommits} Commits
-                    </span>
-                    <span className="flex items-center gap-1">
-                      <svg
-                        className="w-3.5 h-3.5 text-blue-600"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
-                          clipRule="evenodd"
-                        />
-                      </svg>
-                      {selectedUser.totalPRs} PRs
-                    </span>
-                    <span className="flex items-center gap-1">
-                      <svg
-                        className="w-3.5 h-3.5 text-purple-600"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-7-4a1 1 0 11-2 0 1 1 0 012 0zM9 9a1 1 0 000 2v3a1 1 0 001 1h1a1 1 0 100-2v-3a1 1 0 00-1-1H9z"
-                          clipRule="evenodd"
-                        />
-                      </svg>
-                      {selectedUser.totalIssues} Issues
-                    </span>
-                  </div>
-
-                  {/* Man hours rationale */}
-                  {selectedUser.manHoursRationale && (
-                    <div>
-                      <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                        Time Estimate Rationale
-                      </h4>
-                      <p className="text-sm text-gray-600 bg-gray-50 p-3 rounded-lg">
-                        {selectedUser.manHoursRationale}
-                      </p>
-                    </div>
-                  )}
-
-                  {/* Work done */}
-                  {selectedUser.workDone.length > 0 && (
-                    <div>
-                      <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                        Work Completed
-                      </h4>
-                      <ul className="space-y-1">
-                        {selectedUser.workDone.map((item, index) => (
-                          <li
-                            key={index}
-                            className="flex items-start gap-2 text-sm text-gray-700"
-                          >
-                            <span className="text-emerald-500 mt-0.5 font-medium">
-                              ✓
-                            </span>
-                            <span>{item}</span>
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-
-                  {/* Working on */}
-                  {selectedUser.workingOn.length > 0 && (
-                    <div>
-                      <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                        Currently Working On
-                      </h4>
-                      <ul className="space-y-1">
-                        {selectedUser.workingOn.map((item, index) => (
-                          <li
-                            key={index}
-                            className="flex items-start gap-2 text-sm text-gray-700"
-                          >
-                            <span className="text-blue-500 mt-0.5 font-medium">
-                              ⚡
-                            </span>
-                            <span>{item}</span>
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-
-                  {/* Ongoing issues */}
-                  {selectedUser.ongoingIssues.length > 0 && (
-                    <div>
-                      <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                        Ongoing Issues
-                      </h4>
-                      <ul className="space-y-1">
-                        {selectedUser.ongoingIssues.map((item, index) => (
-                          <li
-                            key={index}
-                            className="flex items-start gap-2 text-sm text-gray-700"
-                          >
-                            <span className="text-red-500 mt-0.5 font-medium">
-                              ⚠
-                            </span>
-                            <span>{item}</span>
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-
-                  {/* Daily Timeline */}
-                  {selectedUser.dailyStandups &&
-                    selectedUser.dailyStandups.length > 0 && (
-                      <div>
-                        <h4 className="text-sm font-semibold text-gray-700 mb-3">
-                          Daily Standup
-                        </h4>
-                        <div className="">
-                          {selectedUser.dailyStandups.map((daily, index) => (
-                            <div
-                              key={index}
-                              className="border-l-2 border-gray-200 pl-4 pb-6 last:pb-0"
-                            >
-                              <div className="flex items-center gap-2 mb-2">
-                                <div className="w-2 h-2 bg-indigo-500 rounded-full -ml-5 border-2 border-white"></div>
-                                <h5 className="text-xs font-medium text-gray-900">
-                                  {daily.date
-                                    ? format(
-                                        new Date(daily.date),
-                                        "MMM d, yyyy"
-                                      )
-                                    : `Day ${index + 1}`}
-                                </h5>
-                                <div className="flex items-center gap-2 text-xs text-gray-500">
-                                  <span>{daily.totalCommits}c</span>
-                                  <span>{daily.totalPRs}pr</span>
-                                  <span>{daily.totalIssues}i</span>
-                                  <span className="text-green-600">
-                                    {daily.totalManHoursMin}-
-                                    {daily.totalManHoursMax}h
-                                  </span>
-                                </div>
-                              </div>
-
-                              {daily.summary && (
-                                <p className="text-xs text-gray-600 mb-2 bg-gray-50 p-2 rounded">
-                                  {daily.summary}
-                                </p>
-                              )}
-
-                              <div className="space-y-2">
-                                {daily.workDone.length > 0 && (
-                                  <div>
-                                    <p className="text-xs font-medium text-gray-700 mb-1">
-                                      Completed:
-                                    </p>
-                                    <ul className="space-y-1 ml-2">
-                                      {daily.workDone.map((item, itemIndex) => (
-                                        <li
-                                          key={itemIndex}
-                                          className="flex items-start gap-1 text-xs text-gray-600"
-                                        >
-                                          <span className="text-emerald-500 mt-0.5">
-                                            ✓
-                                          </span>
-                                          <span>{item}</span>
-                                        </li>
-                                      ))}
-                                    </ul>
-                                  </div>
-                                )}
-
-                                {daily.workingOn.length > 0 && (
-                                  <div>
-                                    <p className="text-xs font-medium text-gray-700 mb-1">
-                                      Working on:
-                                    </p>
-                                    <ul className="space-y-1 ml-2">
-                                      {daily.workingOn.map(
-                                        (item, itemIndex) => (
-                                          <li
-                                            key={itemIndex}
-                                            className="flex items-start gap-1 text-xs text-gray-600"
-                                          >
-                                            <span className="text-blue-500 mt-0.5">
-                                              ⚡
-                                            </span>
-                                            <span>{item}</span>
-                                          </li>
-                                        )
-                                      )}
-                                    </ul>
-                                  </div>
-                                )}
-
-                                {daily.ongoingIssues.length > 0 && (
-                                  <div>
-                                    <p className="text-xs font-medium text-gray-700 mb-1">
-                                      Issues:
-                                    </p>
-                                    <ul className="space-y-1 ml-2">
-                                      {daily.ongoingIssues.map(
-                                        (item, itemIndex) => (
-                                          <li
-                                            key={itemIndex}
-                                            className="flex items-start gap-1 text-xs text-gray-600"
-                                          >
-                                            <span className="text-red-500 mt-0.5">
-                                              ⚠
-                                            </span>
-                                            <span>{item}</span>
-                                          </li>
-                                        )
-                                      )}
-                                    </ul>
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                </div>
-              </div>
-            </>
-          ) : (
-            <div className="flex-1 flex items-center justify-center p-8 text-center text-gray-500">
-              <p className="max-w-sm">
-                Select a team member from the ranking to view detailed
-                performance metrics and standups.
-              </p>
-            </div>
-          )}
+          <UserDetailedView selectedUser={selectedUser} />
         </div>
       </div>
 
       {/* Mobile Modal for Detailed View */}
-      {selectedUser && (
-        <div
-          className="lg:hidden fixed inset-0 z-50 flex items-end justify-center"
-          onClick={handleCloseModal}
-        >
-          <div className="absolute inset-0 bg-black opacity-50"></div>
-          <div
-            className={`relative bg-white rounded-t-2xl w-full max-h-[90vh] flex flex-col ${
-              isModalClosing ? "animate-slide-down" : "animate-slide-up"
-            }`}
-            onClick={(e) => e.stopPropagation()}
-          >
-            {/* Modal Header */}
-            <div className="flex-shrink-0 border-b border-gray-200 px-4 py-3 flex items-center justify-between">
-              <div className="flex items-center gap-3 min-w-0 flex-1">
-                <img
-                  src={selectedUser.avatar_url}
-                  alt={selectedUser.username}
-                  className="w-10 h-10 rounded-full flex-shrink-0"
-                />
-                <div className="min-w-0 flex-1">
-                  <h3 className="font-semibold text-gray-900 truncate">
-                    {selectedUser.name || selectedUser.username}
-                  </h3>
-                  {selectedUser.name && (
-                    <p className="text-sm text-gray-500 truncate">
-                      @{selectedUser.username}
-                    </p>
-                  )}
-                </div>
-              </div>
-              <button
-                onClick={handleCloseModal}
-                className="p-2 hover:bg-gray-100 rounded-full transition-colors flex-shrink-0"
-              >
-                <svg
-                  className="w-5 h-5 text-gray-500"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-              </button>
-            </div>
-
-            {/* Modal Content */}
-            <div className="flex-1 overflow-y-auto px-4 py-6 space-y-4">
-              {/* Metrics */}
-              <div className="grid grid-cols-2 gap-3">
-                <div className="bg-gray-50 p-3 rounded-lg text-center">
-                  <div className="text-lg font-semibold text-green-600">
-                    {selectedUser.totalManHoursMin}-
-                    {selectedUser.totalManHoursMax}h
-                  </div>
-                  <div className="text-sm text-gray-500">Man-Hours</div>
-                </div>
-                <div className="bg-gray-50 p-3 rounded-lg text-center">
-                  <div className="text-lg font-semibold text-blue-600">
-                    {selectedUser.avgManHours.toFixed(1)}h
-                  </div>
-                  <div className="text-sm text-gray-500">Average</div>
-                </div>
-                <div className="bg-gray-50 p-3 rounded-lg text-center">
-                  <div className="text-lg font-semibold text-purple-600">
-                    {selectedUser.totalActivities}
-                  </div>
-                  <div className="text-sm text-gray-500">Activities</div>
-                </div>
-                <div className="bg-gray-50 p-3 rounded-lg text-center">
-                  <div className="text-lg font-semibold text-orange-600">
-                    {selectedUser.activeDays}
-                  </div>
-                  <div className="text-sm text-gray-500">Active Days</div>
-                </div>
-              </div>
-
-              {/* Activity breakdown */}
-              <div className="flex items-center justify-center gap-4 text-sm text-gray-600 bg-gray-50 p-3 rounded-lg">
-                <span className="flex items-center gap-1">
-                  <svg
-                    className="w-4 h-4 text-green-600"
-                    fill="currentColor"
-                    viewBox="0 0 20 20"
-                  >
-                    <path
-                      fillRule="evenodd"
-                      d="M2 5a2 2 0 012-2h12a2 2 0 012 2v10a2 2 0 01-2 2H4a2 2 0 01-2-2V5zm3.293 1.293a1 1 0 011.414 0l3 3a1 1 0 010 1.414l-3 3a1 1 0 01-1.414-1.414L7.586 10 5.293 7.707a1 1 0 010-1.414zM11 12a1 1 0 100 2h3a1 1 0 100-2h-3z"
-                      clipRule="evenodd"
-                    />
-                  </svg>
-                  {selectedUser.totalCommits} Commits
-                </span>
-                <span className="flex items-center gap-1">
-                  <svg
-                    className="w-4 h-4 text-blue-600"
-                    fill="currentColor"
-                    viewBox="0 0 20 20"
-                  >
-                    <path
-                      fillRule="evenodd"
-                      d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
-                      clipRule="evenodd"
-                    />
-                  </svg>
-                  {selectedUser.totalPRs} PRs
-                </span>
-                <span className="flex items-center gap-1">
-                  <svg
-                    className="w-4 h-4 text-purple-600"
-                    fill="currentColor"
-                    viewBox="0 0 20 20"
-                  >
-                    <path
-                      fillRule="evenodd"
-                      d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-7-4a1 1 0 11-2 0 1 1 0 012 0zM9 9a1 1 0 000 2v3a1 1 0 001 1h1a1 1 0 100-2v-3a1 1 0 00-1-1H9z"
-                      clipRule="evenodd"
-                    />
-                  </svg>
-                  {selectedUser.totalIssues} Issues
-                </span>
-              </div>
-
-              {/* Man hours rationale */}
-              {selectedUser.manHoursRationale && (
-                <div>
-                  <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                    Time Estimate Rationale
-                  </h4>
-                  <p className="text-sm text-gray-600 bg-gray-50 p-3 rounded-lg">
-                    {selectedUser.manHoursRationale}
-                  </p>
-                </div>
-              )}
-
-              {/* Work done */}
-              {selectedUser.workDone.length > 0 && (
-                <div>
-                  <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                    Work Completed
-                  </h4>
-                  <ul className="space-y-2">
-                    {selectedUser.workDone.map((item, index) => (
-                      <li
-                        key={index}
-                        className="flex items-start gap-2 text-sm text-gray-700"
-                      >
-                        <span className="text-emerald-500 mt-0.5 font-medium">
-                          ✓
-                        </span>
-                        <span>{item}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {/* Working on */}
-              {selectedUser.workingOn.length > 0 && (
-                <div>
-                  <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                    Currently Working On
-                  </h4>
-                  <ul className="space-y-2">
-                    {selectedUser.workingOn.map((item, index) => (
-                      <li
-                        key={index}
-                        className="flex items-start gap-2 text-sm text-gray-700"
-                      >
-                        <span className="text-blue-500 mt-0.5 font-medium">
-                          ⚡
-                        </span>
-                        <span>{item}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {/* Ongoing issues */}
-              {selectedUser.ongoingIssues.length > 0 && (
-                <div>
-                  <h4 className="text-sm font-semibold text-gray-700 mb-2">
-                    Ongoing Issues
-                  </h4>
-                  <ul className="space-y-2">
-                    {selectedUser.ongoingIssues.map((item, index) => (
-                      <li
-                        key={index}
-                        className="flex items-start gap-2 text-sm text-gray-700"
-                      >
-                        <span className="text-red-500 mt-0.5 font-medium">
-                          ⚠
-                        </span>
-                        <span>{item}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {/* Daily Timeline */}
-              {selectedUser.dailyStandups &&
-                selectedUser.dailyStandups.length > 0 && (
-                  <div>
-                    <h4 className="text-sm font-semibold text-gray-700 mb-3">
-                      Daily Standup
-                    </h4>
-                    <div className="">
-                      {selectedUser.dailyStandups.map((daily, index) => (
-                        <div
-                          key={index}
-                          className="border-l-2 border-gray-200 pl-4 pb-6 last:pb-0"
-                        >
-                          <div className="flex items-center gap-2 mb-2">
-                            <div className="w-2 h-2 bg-indigo-500 rounded-full -ml-5 border-2 border-white"></div>
-                            <h5 className="text-sm font-medium text-gray-900">
-                              {daily.date
-                                ? format(new Date(daily.date), "MMM d, yyyy")
-                                : `Day ${index + 1}`}
-                            </h5>
-                          </div>
-
-                          <div className="flex items-center gap-3 text-xs text-gray-500 mb-2">
-                            <span>{daily.totalCommits} commits</span>
-                            <span>{daily.totalPRs} PRs</span>
-                            <span>{daily.totalIssues} issues</span>
-                            <span className="text-green-600">
-                              {daily.totalManHoursMin}-{daily.totalManHoursMax}h
-                            </span>
-                          </div>
-
-                          {daily.summary && (
-                            <p className="text-sm text-gray-600 mb-3 bg-gray-50 p-3 rounded">
-                              {daily.summary}
-                            </p>
-                          )}
-
-                          <div className="space-y-3">
-                            {daily.workDone.length > 0 && (
-                              <div>
-                                <p className="text-sm font-medium text-gray-700 mb-2">
-                                  Completed:
-                                </p>
-                                <ul className="space-y-1 ml-3">
-                                  {daily.workDone.map((item, itemIndex) => (
-                                    <li
-                                      key={itemIndex}
-                                      className="flex items-start gap-2 text-sm text-gray-600"
-                                    >
-                                      <span className="text-emerald-500 mt-0.5">
-                                        ✓
-                                      </span>
-                                      <span>{item}</span>
-                                    </li>
-                                  ))}
-                                </ul>
-                              </div>
-                            )}
-
-                            {daily.workingOn.length > 0 && (
-                              <div>
-                                <p className="text-sm font-medium text-gray-700 mb-2">
-                                  Working on:
-                                </p>
-                                <ul className="space-y-1 ml-3">
-                                  {daily.workingOn.map((item, itemIndex) => (
-                                    <li
-                                      key={itemIndex}
-                                      className="flex items-start gap-2 text-sm text-gray-600"
-                                    >
-                                      <span className="text-blue-500 mt-0.5">
-                                        ⚡
-                                      </span>
-                                      <span>{item}</span>
-                                    </li>
-                                  ))}
-                                </ul>
-                              </div>
-                            )}
-
-                            {daily.ongoingIssues.length > 0 && (
-                              <div>
-                                <p className="text-sm font-medium text-gray-700 mb-2">
-                                  Issues:
-                                </p>
-                                <ul className="space-y-1 ml-3">
-                                  {daily.ongoingIssues.map(
-                                    (item, itemIndex) => (
-                                      <li
-                                        key={itemIndex}
-                                        className="flex items-start gap-2 text-sm text-gray-600"
-                                      >
-                                        <span className="text-red-500 mt-0.5">
-                                          ⚠
-                                        </span>
-                                        <span>{item}</span>
-                                      </li>
-                                    )
-                                  )}
-                                </ul>
-                              </div>
-                            )}
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-            </div>
-          </div>
-        </div>
-      )}
+      <MobileModal
+        selectedUser={selectedUser}
+        isModalClosing={isModalClosing}
+        onClose={handleCloseModal}
+      />
     </div>
   );
 }
